Generate Makefile from a given list of rules (job).

In the (GNU) make jargon, a rule is a sequence of commands to build a result. In this package, a rule should be understood similarly: It is a command or a sequence of commands that optionally produces some files and depends on some other files (such as data files or scripts) or other rules. Moreover, a rule contains a command for cleanup, i.e., for removal of generated files.

Usage

makefile(
  job = list(),
  fileName = NULL,
  makeScript = "Makefile.R",
  vars = NULL,
  all = TRUE,
  tasks = TRUE,
  clean = TRUE,
  makefile = TRUE,
  depends = NULL
)

Arguments

job

A list of rules (i.e., instances of the S3 class rmake.rule - see rule())

fileName

A file to write to. If NULL, the result is returned as a character vector instead of writing to a file.

makeScript

The name of the file that calls this function (used to generate the makefile rule)

vars

A named character vector of shell variables that will be declared in the resulting Makefile (in addition to [defaultVars])

all

TRUE if the all rule should be automatically created and added: the created all rule has dependencies on all the other rules, which causes everything to be built if make all is executed in the shell's command line.

tasks

TRUE if "task" rules should be automatically created and added – see rule() for more details.

clean

TRUE if the clean rule should be automatically created and added

makefile

TRUE if the Makefile rule should be automatically created and added: this rule ensures that any change in the R script that generates the Makefile (i.e., that calls makefile()) triggers the re-generation of the Makefile at the beginning of any build.

depends

A character vector of file names that the makefile generating script depends on

Value

If fileName is NULL, the function returns a character vector with the contents of the Makefile. Otherwise, the content is written to the given fileName.

Details

The makefile() function takes a list of rules (see rule()) and generates a Makefile from them. Additionally, all and clean rules are optionally generated too, which can be executed from the shell by issuing the make all or make clean command, respectively, to build everything or erase all generated files.

If there is a need to group some rules together, it can be done either via dependencies or by using the task mechanism. Each rule may be assigned one or more tasks (see task in rule()). Each task is then created as a standalone rule depending on the assigned rules. That way, executing make task_name will build all rules with the assigned task task_name. By default, all rules are assigned to task all, which allows make all to build everything.

See also

Author

Michal Burda

Examples

# create some jobs
job <- list(
    rRule('dataset.rds', 'preprocess.R', 'dataset.csv'),
    markdownRule('report.pdf', 'report.Rmd', 'dataset.rds'),
    markdownRule('details.pdf', 'details.Rmd', 'dataset.rds'))

# generate Makefile (output as a character vector)
makefile(job)
#>  [1] "# Generated by rmake: do not edit by hand"                                                                               
#>  [2] "SHELL=/bin/sh"                                                                                                           
#>  [3] "R=\"$(R_HOME)/bin$(R_ARCH)/Rscript\""                                                                                    
#>  [4] "RM=rm"                                                                                                                   
#>  [5] "CP=cp"                                                                                                                   
#>  [6] ".ONESHELL:"                                                                                                              
#>  [7] ""                                                                                                                        
#>  [8] ".PHONY: all"                                                                                                             
#>  [9] "all: Makefile dataset.rds report.pdf details.pdf"                                                                        
#> [10] "\t"                                                                                                                      
#> [11] "dataset.rds: preprocess.R dataset.csv"                                                                                   
#> [12] "\t$(R) - <<'EOFrmake'"                                                                                                   
#> [13] "\t{"                                                                                                                     
#> [14] "\t    params <- list(.target = \"dataset.rds\", .script = \"preprocess.R\", .depends = \"dataset.csv\", .task = \"all\")"
#> [15] "\t    source(\"preprocess.R\")"                                                                                          
#> [16] "\t}"                                                                                                                     
#> [17] "\tEOFrmake"                                                                                                              
#> [18] "report.pdf: report.Rmd dataset.rds"                                                                                      
#> [19] "\t$(R) - <<'EOFrmake'"                                                                                                   
#> [20] "\t{"                                                                                                                     
#> [21] "\t    params <- list(.target = \"report.pdf\", .script = \"report.Rmd\", .depends = \"dataset.rds\", .task = \"all\")"   
#> [22] "\t    targets <- list(report.pdf = \"pdf_document\")"                                                                    
#> [23] "\t    for (tg in names(targets)) {"                                                                                      
#> [24] "\t        rmarkdown::render(\"report.Rmd\", output_format = targets[[tg]], output_file = tg)"                            
#> [25] "\t    }"                                                                                                                 
#> [26] "\t}"                                                                                                                     
#> [27] "\tEOFrmake"                                                                                                              
#> [28] "details.pdf: details.Rmd dataset.rds"                                                                                    
#> [29] "\t$(R) - <<'EOFrmake'"                                                                                                   
#> [30] "\t{"                                                                                                                     
#> [31] "\t    params <- list(.target = \"details.pdf\", .script = \"details.Rmd\", .depends = \"dataset.rds\", .task = \"all\")" 
#> [32] "\t    targets <- list(details.pdf = \"pdf_document\")"                                                                   
#> [33] "\t    for (tg in names(targets)) {"                                                                                      
#> [34] "\t        rmarkdown::render(\"details.Rmd\", output_format = targets[[tg]], output_file = tg)"                           
#> [35] "\t    }"                                                                                                                 
#> [36] "\t}"                                                                                                                     
#> [37] "\tEOFrmake"                                                                                                              
#> [38] ".PHONY: clean"                                                                                                           
#> [39] "clean: "                                                                                                                 
#> [40] "\t$(RM) dataset.rds"                                                                                                     
#> [41] "\t$(RM) report.pdf"                                                                                                      
#> [42] "\t$(RM) details.pdf"                                                                                                     
#> [43] "Makefile: Makefile.R"                                                                                                    
#> [44] "\t$(R) - <<'EOFrmake'"                                                                                                   
#> [45] "\t{"                                                                                                                     
#> [46] "\t    params <- list(.target = \"Makefile\", .script = \"Makefile.R\", .depends = NULL, .task = \"all\")"                
#> [47] "\t    source(\"Makefile.R\")"                                                                                            
#> [48] "\t}"                                                                                                                     
#> [49] "\tEOFrmake"                                                                                                              

# generate to file
tmp <- tempdir()
makefile(job, file.path(tmp, "Makefile"))